Python file for model prediction
This file does not use flask. Instead it locally loads the image to be predicted, does the image preprocessing using the keras.preprocessing.image library(instead of using pillow or cv2) and then predicts the result using model.predict. This was created to test if the image prediction works without using the base64 image encoding using JSON. It does. So now, instead of using base64, I will be using formdata instead